~ Data Shop

+++ Assets/_UI/Scripts/ShopDetail.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class ShopDetail : UICanvas
 {
-    public float Id;
+    public int Id;
     private void Start() {
         
     }
